requirements.txt
cmcrameri>=1.9
datasets>=4.2.0
hydra-core>=1.3.2
ipykernel
matplotlib>=3.10
nnsight>=0.5.9
numpy>=2.3
omegaconf
seaborn
pandas>=2.3
polars>=1.33
pytest
scikit-learn==1.7.2
scipy
spacy
omegaconf
torch==2.6.0
tqdm
transformers==4.57.1
bitsandbytes